perf(build): optimize gradle caching and build performance anf fix build errors
- Detect release vs debug builds to avoid cache invalidation on every debug build
- Use static values for debug builds and dynamic git metadata only for releases
- Disable resource optimization (minification, PNG crunching) for faster debug builds
- Increase Gradle daemon heap to 8GB and add separate Kotlin daemon memory settings
- Add worker count optimization comment for multi-core systems
- Remove deprecated gradle.configureondemand property
- Add Quote model for simplified quote sharing and card generation
- Fix iOS encryption service type safety issues (NSData options parameter)
- Fix memcpy call signature in iOS byte array conversion
- Add settings-gradle.lockfile for dependency lock management
- Update local.properties.example with configuration guidance
- Update i18n build configuration

+++ domain/src/iosMain/kotlin/ireader/domain/services/sync/CommonEncryptionService.ios.kt
@@ -48,7 +48,7 @@ actual class CommonEncryptionService actual constructor() : EncryptionService {
         val plaintextBytes = plaintext.encodeToByteArray()
         val encryptedBytes = encryptBytes(plaintextBytes, key)
         
-        return encryptedBytes.toNSData().base64EncodedStringWithOptions(0)
+        return encryptedBytes.toNSData().base64EncodedStringWithOptions(0u)
     }
     
     /**
@@ -59,7 +59,7 @@ actual class CommonEncryptionService actual constructor() : EncryptionService {
             "Key must be $KEY_SIZE_BYTES bytes (256 bits), got ${key.size} bytes"
         }
         
-        val encryptedData = NSData.create(base64EncodedString = ciphertext, options = 0)
+        val encryptedData = NSData.create(base64EncodedString = ciphertext, options = 0u)
             ?: throw IllegalArgumentException("Invalid Base64 string")
         
         val encryptedBytes = encryptedData.toByteArray()
@@ -158,7 +158,7 @@ actual class CommonEncryptionService actual constructor() : EncryptionService {
     private fun NSData.toByteArray(): ByteArray {
         return ByteArray(this.length.toInt()).apply {
             usePinned { pinned ->
-                platform.posix.memcpy(pinned.addressOf(0), this@toByteArray.bytes, this@toByteArray.length.toULong())
+                platform.posix.memcpy(pinned.addressOf(0), this@toByteArray.bytes, this@toByteArray.length)
             }
         }
     }
